backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-simple-history>=3.4,<4.0
django-filter>=23.5
psycopg2-binary>=2.9,<3.0
python-decouple>=3.8
